Implemented multiplayer with tcp
Added a tcp.rs file which handles all tcp interaction according to our agreed upon protocol. It includes a Message type which has all the data needed. It parses these to and from strings and sends these over tcp.
In main.rs, we set up the tcp connection before we start the app. The ip adress and whether we are server or client are provided via cli args. Once the connection is established, we start the app. Then, before we handle the moves, we check if it is our turn to move, if not, we just try to read the move over tcp. If it is our turn, we perform the move and then send the message